1. Add a Messages class that emits direct message events and exposes the sendDirectMessage API. 2. Factor out a NotifierProxy class that Descriptor and Messages inherit from to implement the event emitting functionality including pause/unpause.
